feat: refactor TopRating feature to use MovieRepository and improve movie retrieval logic
