+++ app/src/main/java/com/baghdad/novix/di/RepositoryModule.kt
@@ -6,7 +6,6 @@ import com.baghdad.domain.repository.FavoriteGenreRepository
 import com.baghdad.domain.repository.MovieRepository
 import com.baghdad.domain.repository.RecentlyViewedRepository
 import com.baghdad.domain.repository.SearchRepository
-import com.baghdad.domain.repository.TopRatingRepository
 import com.baghdad.domain.repository.TvShowRepository
 import com.baghdad.repository.ActorRepositoryImpl
 import com.baghdad.repository.EpisodeRepositoryImpl
@@ -14,7 +13,6 @@ import com.baghdad.repository.FavoriteGenreRepositoryImpl
 import com.baghdad.repository.MovieRepositoryImpl
 import com.baghdad.repository.RecentlyViewedRepositoryImpl
 import com.baghdad.repository.SearchRepositoryImpl
-import com.baghdad.repository.TopRatingRepositoryImpl
 import com.baghdad.repository.TvShowRepositoryImpl
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
@@ -28,5 +26,4 @@ val repositoryModule = module {
     singleOf(::ActorRepositoryImpl) { bind<ActorRepository>() }
     singleOf(::EpisodeRepositoryImpl) { bind<EpisodeRepository>() }
     singleOf(::FavoriteGenreRepositoryImpl) { bind<FavoriteGenreRepository>() }
-    singleOf(::TopRatingRepositoryImpl) { bind<TopRatingRepository>() }
 }
