+++ domain/src/main/java/com/baghdad/domain/repository/MovieRepository.kt
@@ -1,5 +1,6 @@
 package com.baghdad.domain.repository
 
+import com.baghdad.domain.model.PagedResult
 import com.baghdad.entity.media.Genre
 import com.baghdad.entity.media.Movie
 import com.baghdad.entity.media.Review
@@ -13,5 +14,5 @@ interface MovieRepository {
     suspend fun getMoviesByGenre(genreId: Long, page: Int): List<Movie>
     suspend fun getMovieReviews(movieId: Long): List<Review>
     suspend fun getMovieImages(movieId: Long): List<String>
-    suspend fun getTopRatedMovies(page: Int): List<Movie>
+    suspend fun getTopRatedMovies(page: Int): PagedResult<Movie>
 }
