--- domain/src/main/java/com/baghdad/domain/repository/TopRatingRepository.kt
@@ -1,8 +1,0 @@
-package com.baghdad.domain.repository
-
-import com.baghdad.domain.model.PagedResult
-import com.baghdad.entity.media.Movie
-
-interface TopRatingRepository {
-    suspend fun getTopRatedMovies(page: Int): PagedResult<Movie>
-}
