+++ domain/src/main/java/com/baghdad/domain/usecase/movie/GetMovieTopRatingUseCase.kt
@@ -1,19 +1,17 @@
 package com.baghdad.domain.usecase.movie
 
-import android.util.Log
 import com.baghdad.domain.model.PagedResult
-import com.baghdad.domain.repository.TopRatingRepository
+import com.baghdad.domain.repository.MovieRepository
 import com.baghdad.entity.media.Movie
 
 class GetMovieTopRatingUseCase(
-    private val topRatingRepository: TopRatingRepository,
+    private val topRatingRepository: MovieRepository,
 ) {
     suspend operator fun invoke(
         page: Int,
         genreId: Long
     ): PagedResult<Movie> {
         val result = topRatingRepository.getTopRatedMovies(page)
-
         val filteredMovies = if (genreId != 0L) {
             result.data.filter { movie ->
                 movie.genres.any { genre -> genre.id == genreId }
@@ -21,16 +19,6 @@ class GetMovieTopRatingUseCase(
         } else {
             result.data
         }
-
-        Log.d(
-            "GetMovieTopRatingUseCase", "Filtered ${
-                filteredMovies.map {
-                    it.genres.size
-                }
-            }"
-        )
-        Log.d("GetMovieTopRatingUseCase", "Filtered Id is $genreId")
-
         return result.copy(data = filteredMovies)
     }
