+++ local_datasource/src/main/java/com/baghdad/local_datasource/LocalTopRatingDataSourceImpl.kt
@@ -1,10 +1,10 @@
 package com.baghdad.local_datasource
 
-import android.util.Log
 import com.baghdad.local_datasource.roomDB.dao.GenreDao
 import com.baghdad.local_datasource.roomDB.dao.TopRatedDao
 import com.baghdad.local_datasource.roomDB.entity.toDto
 import com.baghdad.local_datasource.roomDB.entity.toLocalDtos
+import com.baghdad.local_datasource.roomDB.errorHandler.executeWithErrorHandling
 import com.baghdad.local_datasource.util.calculatePageOffset
 import com.baghdad.repository.datasource.local.LocalTopRatingDataSource
 import com.baghdad.repository.logger.Logger
@@ -18,8 +18,10 @@ class LocalTopRatingDataSourceImpl(
 ) : LocalTopRatingDataSource {
 
     override suspend fun getMovieGenre(language: String): List<GenreDto> {
-        return genreDao.getAllGenres().filter { it.type == GenreDto.GenreType.MOVIE.name }
-            .map { it.toDto() }
+        return executeWithErrorHandling(logger = logger) {
+            genreDao.getAllGenres().filter { it.type == GenreDto.GenreType.MOVIE.name }
+                .map { it.toDto() }
+        }
     }
 
     override suspend fun getTopRatedMovies(
@@ -27,19 +29,19 @@ class LocalTopRatingDataSourceImpl(
         pageSize: Int,
     ): List<MovieDto> {
         val offset = calculatePageOffset(page, pageSize)
-        val xxx= topRatedDao.getTopRatedMoves(pageSize, offset).map {
-            val genresDto = it.genres.map { genreId ->
-                genreDao.getGenreById(genreId).toDto()
+        return executeWithErrorHandling(logger = logger) {
+            topRatedDao.getTopRatedMoves(pageSize, offset).map {
+                val genresDto = it.genres.map { genreId ->
+                    genreDao.getGenreById(genreId).toDto()
+                }
+                it.toDto(genresDto)
             }
-            it.toDto(genresDto)
         }
-        Log.d("LocalTopRatingDataSourceImpl", "📦 Retrieved ${xxx} top-rated movies for page=$page, offset=$offset")
-        return xxx
     }
 
     override suspend fun saveTopRatedMovies(movieData: List<MovieDto>) {
-       // Log.d("LocalTopRatingDataSourceImpl", "💾 Saving ${movieData} top-rated movies")
-        topRatedDao.upsertMovies(movieData.map { it.toLocalDtos() })
-        //Log.d("LocalTopRatingDataSourceImpl", "💾 Saving ${movieData.map { it.toLocalDtos() }} top-rated movies")
+        executeWithErrorHandling(logger = logger) {
+            topRatedDao.upsertMovies(movieData.map { it.toLocalDtos() })
+        }
     }
 }
