+++ repository/src/main/java/com/baghdad/repository/MovieRepositoryImpl.kt
@@ -1,20 +1,28 @@
 package com.baghdad.repository
 
-import android.util.Log
+import com.baghdad.domain.model.PagedResult
 import com.baghdad.domain.repository.MovieRepository
 import com.baghdad.entity.media.Genre
 import com.baghdad.entity.media.Movie
 import com.baghdad.entity.media.Review
 import com.baghdad.entity.person.CastMember
+import com.baghdad.repository.datasource.local.LocalGenreDataSource
+import com.baghdad.repository.datasource.local.LocalTopRatingDataSource
 import com.baghdad.repository.datasource.remote.RemoteGenreDataSource
 import com.baghdad.repository.datasource.remote.RemoteMovieDataSource
+import com.baghdad.repository.datasource.remote.RemoteTopRatingDataSource
 import com.baghdad.repository.mapper.toEntity
+import com.baghdad.repository.model.MovieDto
 import com.baghdad.repository.util.executeSafely
+import com.baghdad.repository.util.getPagedSafely
 import java.util.Locale
 
 class MovieRepositoryImpl(
-    val remoteGenreDataSource: RemoteGenreDataSource,
-    val remoteMovieDataSource: RemoteMovieDataSource,
+    private val remoteGenreDataSource: RemoteGenreDataSource,
+    private val localGenreDataSource: LocalGenreDataSource,
+    private val remoteMovieDataSource: RemoteMovieDataSource,
+    private val remoteTopRatingDataSource: RemoteTopRatingDataSource,
+    private val localTopRatingDataSource: LocalTopRatingDataSource,
 ) : MovieRepository {
     override suspend fun getGenres(): List<Genre> {
         return executeSafely {
@@ -35,7 +43,6 @@ class MovieRepositoryImpl(
     override suspend fun getMovieDetails(movieId: Long): Movie {
         return executeSafely {
             val movieTrailer = remoteMovieDataSource.getMovieTrailer(movieId)
-            Log.d("MovieRepositoryImpl", "getMovieDetails: $movieTrailer")
             remoteMovieDataSource.getMovieDetails(movieId).toEntity()
                 .copy(trailerURL = movieTrailer)
         }
@@ -61,13 +68,11 @@ class MovieRepositoryImpl(
     }
 
     override suspend fun getMovieReviews(movieId: Long): List<Review> {
-        Log.d("MovieRepositoryImpl", "getMovieReviews: $movieId")
         val result =  executeSafely {
             remoteMovieDataSource.getMovieReviews(movieId).map {
                 it.toEntity()
             }
         }
-        Log.d("MovieRepositoryImpl", "getMovieReviews: $result")
         return result
     }
 
@@ -77,11 +82,34 @@ class MovieRepositoryImpl(
         }
     }
 
-    override suspend fun getTopRatedMovies(page: Int): List<Movie> {
-        return executeSafely {
-            remoteMovieDataSource.getTopRatedMovies(page).map {
-                it.toEntity()
+    override suspend fun getTopRatedMovies(page: Int): PagedResult<Movie> {
+        return getPagedSafely(
+
+            page = page,
+            pageSize = 20,
+            mapToEntity = MovieDto::toEntity,
+            onStart = {  },
+            getCachedPage = { page, pageSize ->
+                localTopRatingDataSource.getTopRatedMovies(page, pageSize)
+            },
+
+            getRemoteData = { page, _ ->
+                updateGenreCache()
+                remoteTopRatingDataSource.getTopRatedMovies(page)
+
+            },
+
+            cacheData = { data ->
+                localTopRatingDataSource.saveTopRatedMovies(data)
             }
-        }
+
+        )
+
+    }
+
+    private suspend fun updateGenreCache() {
+        val lang = Locale.getDefault().language
+        val movieGenres = remoteGenreDataSource.getMovieGenre(lang)
+        movieGenres.forEach { localGenreDataSource.addGenre(it) }
     }
 }
